Fix PHPStan errors and agent bugs
Bug fixes:
- Fix AgentFinished event reporting incorrect status (in_progress → completed)
- Fix duplicate ContinuationEvaluated events in StepByStep
- Fix AgentBuilder $maxRetries not being used (now passes to ToolCallingDriver)
- Fix AbstractAgentDefinition::build() returning stale agent without event handler

PHPStan fixes:
- Add callable signatures to ContentParts and MessageList map/reduce/filter
- Add Closure signature to MockTool and IdempotencyMiddleware
- Add class-string annotations to SchemaDefinition
- Add array type annotations to Task and TaskList
- Fix return type in TodoReminderProcessor::stepsSinceTodo()
- Fix ReActDriver::withCachedContext() to use accessor methods
- Fix DeterministicDriver array key types with array_values()
- Remove redundant int casts in retry policies
- Add Agent::isStateChanged() to prevent extra iterator yields on status change
- Add phpstan.neon ignores for trait property and forward-compatibility checks

+++ packages/http-client/src/Middleware/RetryPolicy.php
@@ -63,9 +63,9 @@ final readonly class RetryPolicy
         $capped = min($base, $this->maxDelayMs);
 
         $delay = match ($this->jitter) {
-            'none' => (int) $capped,
+            'none' => $capped,
             'equal' => (int) ($capped / 2 + random_int(0, (int) ($capped / 2))),
-            default => (int) random_int(0, (int) $capped),
+            default => random_int(0, $capped),
         };
 
         if ($this->respectRetryAfter && $response !== null) {
